fixed some warnings and typos, cosmetics
git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@5063 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/SubPic/DX9SubPic.cpp
@@ -149,7 +149,7 @@ STDMETHODIMP CDX9SubPic::ClearDirtyRect(DWORD color)
 
 		if (spd.bpp == 16) {
 			while (h-- > 0) {
-				memsetw(ptr, color, 2 * m_rcDirty.Width());
+				memsetw(ptr, (unsigned short)color, 2 * m_rcDirty.Width());
 				ptr += spd.pitch;
 			}
 		} else if (spd.bpp == 32) {
