+++ src/filters/source/ShoutcastSource/ShoutcastSource.cpp
@@ -595,7 +595,7 @@ int CShoutcastStream::CShoutcastSocket::Receive(void* lpBuf, int nBufLen, int nF
 
 				TRACE(_T("found possible length byte: %d, skipping %d bytes\n"), *p, 1 + *p*16);
 				p += 1 + *p*16;
-				len = (p0 + len) - p;
+				len = (int)(p0 + len - p);
 				TRACE(_T("returning the remaining bytes in the packet: %d\n"), len);
 				if (len <= 0) {
 					TRACE(_T("nothing to return, reading a bit more in\n"));
